+++ src/Service/Base/BaseService.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * (c) Meritoo.pl, http://www.meritoo.pl
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Meritoo\CommonBundle\Service\Base;
+
+/**
+ * Groundwork of Symfony's service.
+ * Basic methods and functionality.
+ *
+ * @author    Meritoo <github@meritoo.pl>
+ * @copyright Meritoo <http://www.meritoo.pl>
+ */
+abstract class BaseService
+{
+}