+++ tests/Service/Base/BaseService/RealService.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * (c) Meritoo.pl, http://www.meritoo.pl
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Meritoo\Test\CommonBundle\Service\Base\BaseService;
+
+use Meritoo\CommonBundle\Service\Base\BaseService;
+
+/**
+ * Real Service used as instance of BaseService while running tests
+ *
+ * @author    Meritoo <github@meritoo.pl>
+ * @copyright Meritoo <http://www.meritoo.pl>
+ */
+class RealService extends BaseService
+{
+}